backend/requirements.txt
# ============================================
# BACKEND REQUIREMENTS
# API Server cho Flutter App
# ============================================

# === Web Framework ===
fastapi==0.109.0                # REST API framework
uvicorn[standard]==0.27.0       # ASGI server
python-multipart==0.0.6         # File upload support
pydantic==2.5.3                 # Data validation

# === Database (PostgreSQL) ===
psycopg2-binary==2.9.9          # PostgreSQL adapter
sqlalchemy==2.0.25              # ORM
alembic==1.13.1                 # Database migrations

# === AI/Computer Vision ===
ultralytics==8.0.220            # YOLOv8
opencv-python==4.8.1.78         # OpenCV
torch==2.1.0                    # PyTorch
torchvision==0.16.0             # PyTorch Vision
Pillow==10.1.0                  # Image processing
scikit-learn==1.3.2             # K-Means clustering
scikit-image==0.22.0            # Image processing
numpy==1.24.3                   # Numerical computing

# === Authentication (Optional) ===
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4          # Password hashing
python-multipart==0.0.6         # Form data

# === Utils ===
python-dotenv==1.0.0            # Environment variables
aiofiles==23.2.1                # Async file operations

# === Machine Learning (Optional) ===
# tensorflow==2.15.0            # Uncomment nếu cần
# keras==2.15.0                 # Uncomment nếu cần
